gzfile() only if capability ..
git-svn-id: https://svn.r-project.org/R/trunk@15631 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ tests/reg-tests-1.R
@@ -157,13 +157,14 @@ try(eigen(m))
 ## segfaults on 1.2.2
 
 ## 1.3.0 had poor compression on gzfile() with lots of small pieces.
-zz <- gzfile("t1.gz", "w")
-write(1:1000, zz)
-close(zz)
-(sz <- file.info("t1.gz")$size)
-unlink("t1.gz")
-stopifnot(sz < 2000)
-
+if (capabilities("libz")) {
+    zz <- gzfile("t1.gz", "w")
+    write(1:1000, zz)
+    close(zz)
+    (sz <- file.info("t1.gz")$size)
+    unlink("t1.gz")
+    stopifnot(sz < 2000)
+}
 ## PR 1010: plot.mts (type="p") was broken in 1.3.0 and this call failed.
 plot(ts(matrix(runif(10), ncol = 2)), type = "p")
